Marginally improve the access spread in BlazePoolVirtualThreadSafeTap
This should decrease the likelihood of two different threads converging on the same stripe indicies.
